requirements.txt
aioredis
aiosmtplib
alembic
anyio
async-timeout
bcrypt
certifi
cffi
click
colorama
fastapi
greenlet
h11
httpcore
httptools
httpx
idna
Jinja2
Mako
MarkupSafe
mysql-connector-python
mysqlclient
passlib
protobuf
pycparser
pydantic
PyJWT
PyMySQL
python-dotenv
python-multipart
PyYAML
rfc3986
six
sniffio
SQLAlchemy
starlette
typing_extensions
uvicorn
